+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * added `synchronize_on_attribute` for `Monitor`
 * added `FutureCollection`
+* added `put_many` to `PeekableQueue`